[refactor #146] separate welcomeView from chatContent, init css file
