+++ webview-ui/src/components/ChatContent.tsx
@@ -4,9 +4,7 @@ import { VscTrash } from "react-icons/vsc";
 import React, { useEffect, useRef, useState } from "react";
 import { Message } from "../types/messageStoreTypes";
 import { AiOutlineLoading3Quarters } from "react-icons/ai";
-import { RiChatNewLine } from "react-icons/ri";
-import { GrHistory } from "react-icons/gr";
-import getOs from "../utilities/getOs";
+import WelcomeView from "./welcome/WelcomeView";
 
 function ChatContent() {
   const { messages, updateMessage, addMessage, loadMessages, setTimestamp } = useMessagesStore();
@@ -49,8 +47,6 @@ function ChatContent() {
     messageEndRef.current?.scrollIntoView({ behavior: "smooth" }); // 자신이 호출된 요소가 사용자에게 표시되도록 상위 컨테이너를 스크롤
   }, [messages]);
 
-  const os = getOs();
-
   return (
     <main id="chat-container">
       {messages && messages.length > 0 ? (
@@ -66,46 +62,7 @@ function ChatContent() {
           ))}
         </div>
       ) : (
-        <div className="description">
-          <p className="how-to-start-title">
-            <strong>How to Start</strong>
-          </p>
-          <ol className="steps">
-            <li>1. Open the source code file you are working on.</li>
-            <li>2. Enter the problem definition you want to solve.</li>
-            <li>3. Enter the process of solving the problem step by step.</li>
-            {os === "windows" ? (
-              <li>
-                4. Press <kbd>ctrl + ⏎</kbd> to run.
-              </li>
-            ) : (
-              <li>
-                4. Press <kbd>⌘ + ⏎</kbd> to run.
-              </li>
-            )}
-            <li>5. After the guiding questions are generated, ask additional questions if needed</li>
-            {os === "windows" ? (
-              <li>
-                6. Press <kbd>ctrl + ⏎</kbd> again to run.
-              </li>
-            ) : (
-              <li>
-                6. Press <kbd>⌘ + ⏎</kbd> again to run.
-              </li>
-            )}
-          </ol>
-          <p className="additional-info">
-            <strong className="spacing1">
-              <RiChatNewLine />
-            </strong>{" "}
-            <span className="spacing1">New Chat</span>
-            <br />
-            <strong className="spacing2">
-              <GrHistory />
-            </strong>{" "}
-            <span className="spacing2">Chat History</span>
-          </p>
-        </div>
+        <WelcomeView />
       )}
       {loading ? (
         <div className="loading">
